Further dialogue improvements
Made it possible to have several lines in one dialogue message, it took a while to get that working! Also the voices are no longer in use when you pickup the phone, I will just change it so that it is the same voice as in the intro and tutorial. Next up is even more improvements to the dialogue! woooo