More Menu Changes
Oh My god I hate menus they are so enoying omg I hate it so much. The fade out is making everything so horrible to code why am I doing this to myself, there is still a glitch where you can spam the restart or main menu button in the pause screen and it will load several instances of the same scene and just softlock the game but eh who is going to notice it. Otherwise everything should be working fine I'll have a last bug check before releasing the demo but everthing should be about done woogit add .